Deletes unnecesary menuLinks mixin
